feat!(sdk-rs): remove escrow/marketplace/events/jobs/games/artifacts namespaces
