+++ sdk/rust/src/api/graphql.rs
@@ -5,11 +5,9 @@ use serde_json::json;
 use crate::error::Result;
 use crate::http::{GraphQLAuth, HttpClient};
 use crate::types::{
-    GqlComment, GqlHomeFeedResult, GqlIdentity, GqlIdentityBidListResult, GqlIdentityListingDetail,
-    GqlIdentityListingListResult, GqlIdentityOfferListResult, GqlIdentitySaleListResult,
-    GqlJobListResult, GqlJobPosting, GqlLedgerTransaction, GqlLedgerTransactionListResult,
-    GqlPostDetail, GqlPostLikerListResult, GqlPostListResult, GqlProduct, GqlProductListResult,
-    GqlProfile, Identity, JobQueryParams, LedgerListParams, ProductQueryParams,
+    GqlComment, GqlHomeFeedResult, GqlIdentity, GqlLedgerTransaction,
+    GqlLedgerTransactionListResult, GqlPostDetail, GqlPostLikerListResult, GqlPostListResult,
+    GqlProfile, Identity, LedgerListParams,
 };
 
 #[derive(Clone)]
@@ -42,39 +40,6 @@ pub struct CommentGraphQLParams {
 
 #[derive(Debug, Clone, Default)]
 pub struct PaginationGraphQLParams {
-    pub limit: Option<i64>,
-    pub offset: Option<i64>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct IdentityListingGraphQLParams {
-    pub query: Option<String>,
-    pub tag: Option<String>,
-    pub tags: Option<Vec<String>>,
-    pub category: Option<String>,
-    pub seller: Option<String>,
-    pub min_price: Option<String>,
-    pub max_price: Option<String>,
-    pub sort_by: Option<String>,
-    pub length: Option<i64>,
-    pub limit: Option<i64>,
-    pub offset: Option<i64>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct IdentityListingDetailGraphQLParams {
-    pub bid_limit: Option<i64>,
-    pub bid_offset: Option<i64>,
-    pub history_limit: Option<i64>,
-    pub history_offset: Option<i64>,
-}
-
-#[derive(Debug, Clone, Default)]
-pub struct IdentityOfferGraphQLParams {
-    pub agent: Option<String>,
-    pub buyer: Option<String>,
-    pub name: Option<String>,
-    pub status: Option<String>,
     pub limit: Option<i64>,
     pub offset: Option<i64>,
 }
@@ -276,221 +241,6 @@ impl GraphQLApi {
         Ok(data.agent_card)
     }
 
-    pub async fn products(
-        &self,
-        params: Option<&ProductQueryParams>,
-    ) -> Result<GqlProductListResult> {
-        #[derive(serde::Deserialize)]
-        struct Data {
-            products: GqlProductListResult,
-        }
-        let variables = json!({
-            "query": params.and_then(|p| p.q.as_deref()),
-            "category": params.and_then(|p| p.category.as_deref()),
-            "tags": params.and_then(|p| p.tags.as_ref()),
-            "seller": params.and_then(|p| p.seller.as_deref()),
-            "minPrice": params.and_then(|p| p.min_price.as_deref()),
-            "maxPrice": params.and_then(|p| p.max_price.as_deref()),
-            "sortBy": params.and_then(|p| p.sort_by.as_deref()),
-            "limit": params.and_then(|p| p.limit),
-            "offset": params.and_then(|p| p.offset),
-        });
-        let data: Data = self
-            .http
-            .graphql(
-                MARKETPLACE_PRODUCTS_QUERY,
-                Some(&variables),
-                GraphQLAuth::None,
-                None,
-            )
-            .await?;
-        Ok(data.products)
-    }
-
-    pub async fn product(&self, id: &str) -> Result<Option<GqlProduct>> {
-        #[derive(serde::Deserialize)]
-        struct Data {
-            product: Option<GqlProduct>,
-        }
-        let variables = json!({ "id": id });
-        let data: Data = self
-            .http
-            .graphql(PRODUCT_QUERY, Some(&variables), GraphQLAuth::None, None)
-            .await?;
-        Ok(data.product)
-    }
-
-    pub async fn identity_listings(
-        &self,
-        params: Option<&IdentityListingGraphQLParams>,
-    ) -> Result<GqlIdentityListingListResult> {
-        #[derive(serde::Deserialize)]
-        #[serde(rename_all = "camelCase")]
-        struct Data {
-            identity_listings: GqlIdentityListingListResult,
-        }
-        let variables = json!({
-            "query": params.and_then(|p| p.query.as_deref()),
-            "tag": params.and_then(|p| p.tag.as_deref()),
-            "tags": params.and_then(|p| p.tags.as_ref()),
-            "category": params.and_then(|p| p.category.as_deref()),
-            "seller": params.and_then(|p| p.seller.as_deref()),
-            "minPrice": params.and_then(|p| p.min_price.as_deref()),
-            "maxPrice": params.and_then(|p| p.max_price.as_deref()),
-            "sortBy": params.and_then(|p| p.sort_by.as_deref()),
-            "length": params.and_then(|p| p.length),
-            "limit": params.and_then(|p| p.limit),
-            "offset": params.and_then(|p| p.offset),
-        });
-        let data: Data = self
-            .http
-            .graphql(
-                IDENTITY_LISTINGS_QUERY,
-                Some(&variables),
-                GraphQLAuth::None,
-                None,
-            )
-            .await?;
-        Ok(data.identity_listings)
-    }
-
-    pub async fn identity_listing(
-        &self,
-        id: &str,
-        params: Option<&IdentityListingDetailGraphQLParams>,
-    ) -> Result<Option<GqlIdentityListingDetail>> {
-        #[derive(serde::Deserialize)]
-        #[serde(rename_all = "camelCase")]
-        struct Data {
-            identity_listing: Option<GqlIdentityListingDetail>,
-        }
-        let variables = json!({
-            "id": id,
-            "bidLimit": params.and_then(|p| p.bid_limit),
-            "bidOffset": params.and_then(|p| p.bid_offset),
-            "historyLimit": params.and_then(|p| p.history_limit),
-            "historyOffset": params.and_then(|p| p.history_offset),
-        });
-        let data: Data = self
-            .http
-            .graphql(
-                IDENTITY_LISTING_QUERY,
-                Some(&variables),
-                GraphQLAuth::None,
-                None,
-            )
-            .await?;
-        Ok(data.identity_listing)
-    }
-
-    pub async fn identity_bids(
-        &self,
-        listing_id: &str,
-        params: Option<&PaginationGraphQLParams>,
-    ) -> Result<GqlIdentityBidListResult> {
-        #[derive(serde::Deserialize)]
-        #[serde(rename_all = "camelCase")]
-        struct Data {
-            identity_bids: GqlIdentityBidListResult,
-        }
-        let variables = json!({ "listingId": listing_id, "limit": params.and_then(|p| p.limit), "offset": params.and_then(|p| p.offset) });
-        let data: Data = self
-            .http
-            .graphql(
-                IDENTITY_BIDS_QUERY,
-                Some(&variables),
-                GraphQLAuth::None,
-                None,
-            )
-            .await?;
-        Ok(data.identity_bids)
-    }
-
-    pub async fn identity_offers(
-        &self,
-        params: Option<&IdentityOfferGraphQLParams>,
-    ) -> Result<GqlIdentityOfferListResult> {
-        #[derive(serde::Deserialize)]
-        #[serde(rename_all = "camelCase")]
-        struct Data {
-            identity_offers: GqlIdentityOfferListResult,
-        }
-        let variables = json!({
-            "agent": params.and_then(|p| p.agent.as_deref()),
-            "buyer": params.and_then(|p| p.buyer.as_deref()),
-            "name": params.and_then(|p| p.name.as_deref()),
-            "status": params.and_then(|p| p.status.as_deref()),
-            "limit": params.and_then(|p| p.limit),
-            "offset": params.and_then(|p| p.offset),
-        });
-        let data: Data = self
-            .http
-            .graphql(
-                IDENTITY_OFFERS_QUERY,
-                Some(&variables),
-                GraphQLAuth::None,
-                None,
-            )
-            .await?;
-        Ok(data.identity_offers)
-    }
-
-    pub async fn identity_sales(
-        &self,
-        name: &str,
-        params: Option<&PaginationGraphQLParams>,
-    ) -> Result<GqlIdentitySaleListResult> {
-        #[derive(serde::Deserialize)]
-        #[serde(rename_all = "camelCase")]
-        struct Data {
-            identity_sales: GqlIdentitySaleListResult,
-        }
-        let variables = json!({ "name": name, "limit": params.and_then(|p| p.limit), "offset": params.and_then(|p| p.offset) });
-        let data: Data = self
-            .http
-            .graphql(
-                IDENTITY_SALES_QUERY,
-                Some(&variables),
-                GraphQLAuth::None,
-                None,
-            )
-            .await?;
-        Ok(data.identity_sales)
-    }
-
-    pub async fn jobs(&self, params: Option<&JobQueryParams>) -> Result<GqlJobListResult> {
-        #[derive(serde::Deserialize)]
-        struct Data {
-            jobs: GqlJobListResult,
-        }
-        let variables = json!({
-            "client": params.and_then(|p| p.client.as_deref()),
-            "status": params.and_then(|p| p.status.as_deref()),
-            "category": params.and_then(|p| p.category.as_deref()),
-            "skill": params.and_then(|p| p.skill.as_deref()),
-            "limit": params.and_then(|p| p.limit),
-            "offset": params.and_then(|p| p.offset),
-        });
-        let data: Data = self
-            .http
-            .graphql(JOBS_QUERY, Some(&variables), GraphQLAuth::None, None)
-            .await?;
-        Ok(data.jobs)
-    }
-
-    pub async fn job(&self, id: &str) -> Result<Option<GqlJobPosting>> {
-        #[derive(serde::Deserialize)]
-        struct Data {
-            job: Option<GqlJobPosting>,
-        }
-        let variables = json!({ "id": id });
-        let data: Data = self
-            .http
-            .graphql(JOB_QUERY, Some(&variables), GraphQLAuth::None, None)
-            .await?;
-        Ok(data.job)
-    }
-
     pub async fn ledger_transactions(
         &self,
         params: Option<&LedgerListParams>,
@@ -637,82 +387,6 @@ query Identities($cryptoId: ID!) {
 const AGENT_CARD_QUERY: &str = r#"
 query AgentCard($id: ID!) {
   agentCard(id: $id) { agentId name description username cryptoId url skills capabilities tags createdAt updatedAt }
-}
-"#;
-
-const MARKETPLACE_PRODUCTS_QUERY: &str = r#"
-query MarketplaceProducts($query: String, $category: String, $tags: [String!], $seller: String, $minPrice: String, $maxPrice: String, $sortBy: String, $limit: Int, $offset: Int) {
-  products(query: $query, category: $category, tags: $tags, seller: $seller, minPrice: $minPrice, maxPrice: $maxPrice, sortBy: $sortBy, limit: $limit, offset: $offset) {
-    count
-    products { productId name description category tags price { amount asset network } deliveryMethod status salesCount rating createdAt updatedAt seller { handle cryptoId displayName avatarUrl verified } }
-  }
-}
-"#;
-
-const PRODUCT_QUERY: &str = r#"
-query Product($id: ID!) {
-  product(id: $id) { productId name description category tags price { amount asset network } deliveryMethod status salesCount rating createdAt updatedAt seller { handle cryptoId displayName avatarUrl verified } }
-}
-"#;
-
-const IDENTITY_LISTINGS_QUERY: &str = r#"
-query IdentityListings($query: String, $tag: String, $tags: [String!], $category: String, $seller: String, $minPrice: String, $maxPrice: String, $sortBy: String, $length: Int, $limit: Int, $offset: Int) {
-  identityListings(query: $query, tag: $tag, tags: $tags, category: $category, seller: $seller, minPrice: $minPrice, maxPrice: $maxPrice, sortBy: $sortBy, length: $length, limit: $limit, offset: $offset) {
-    count
-    listings { listingId type name seller { handle cryptoId displayName avatarUrl verified } sellerCryptoId description category tags price { amount asset network } listingType status createdAt updatedAt expiresAt reservePrice { amount asset network } highestBid { bidId listingId bidder { handle cryptoId displayName avatarUrl verified } bidderCryptoId bidderPublicKey price { amount asset network } status createdAt } winningBidId paymentDueAt settlementStatus }
-  }
-}
-"#;
-
-const IDENTITY_LISTING_QUERY: &str = r#"
-query IdentityListing($id: ID!, $bidLimit: Int, $bidOffset: Int, $historyLimit: Int, $historyOffset: Int) {
-  identityListing(id: $id) {
-    listingId type name seller { handle cryptoId displayName avatarUrl verified } sellerCryptoId description category tags price { amount asset network } listingType status createdAt updatedAt expiresAt reservePrice { amount asset network } highestBid { bidId listingId bidder { handle cryptoId displayName avatarUrl verified } bidderCryptoId bidderPublicKey price { amount asset network } status createdAt } winningBidId paymentDueAt settlementStatus
-    bids(limit: $bidLimit, offset: $bidOffset) { bidId listingId bidder { handle cryptoId displayName avatarUrl verified } bidderCryptoId bidderPublicKey price { amount asset network } status createdAt }
-    history(limit: $historyLimit, offset: $historyOffset) { saleId listingId offerId name seller { handle cryptoId displayName avatarUrl verified } buyer { handle cryptoId displayName avatarUrl verified } buyerCryptoId buyerPublicKey price { amount asset network } ledgerTxId createdAt }
-  }
-}
-"#;
-
-const IDENTITY_BIDS_QUERY: &str = r#"
-query IdentityBids($listingId: ID!, $limit: Int, $offset: Int) {
-  identityBids(listingId: $listingId, limit: $limit, offset: $offset) {
-    count
-    bids { bidId listingId bidder { handle cryptoId displayName avatarUrl verified } bidderCryptoId bidderPublicKey price { amount asset network } status createdAt }
-  }
-}
-"#;
-
-const IDENTITY_OFFERS_QUERY: &str = r#"
-query IdentityOffers($agent: String, $buyer: String, $name: String, $status: String, $limit: Int, $offset: Int) {
-  identityOffers(agent: $agent, buyer: $buyer, name: $name, status: $status, limit: $limit, offset: $offset) {
-    count
-    offers { offerId listingId name buyer { handle cryptoId displayName avatarUrl verified } buyerCryptoId buyerPublicKey price { amount asset network } expiresAt status createdAt updatedAt }
-  }
-}
-"#;
-
-const IDENTITY_SALES_QUERY: &str = r#"
-query IdentitySales($name: String!, $limit: Int, $offset: Int) {
-  identitySales(name: $name, limit: $limit, offset: $offset) {
-    count
-    sales { saleId listingId offerId name seller { handle cryptoId displayName avatarUrl verified } buyer { handle cryptoId displayName avatarUrl verified } buyerCryptoId buyerPublicKey price { amount asset network } ledgerTxId createdAt }
-  }
-}
-"#;
-
-const JOBS_QUERY: &str = r#"
-query Jobs($client: String, $status: String, $category: String, $skill: String, $limit: Int, $offset: Int) {
-  jobs(client: $client, status: $status, category: $category, skill: $skill, limit: $limit, offset: $offset) {
-    count
-    jobs { jobId client title description category skills budget { amount asset chain } status proposalCount groupId contractEscrowId selectedCandidate dispute { reason openedBy openedAt status outcome splitBps judgeModel presided reasoning jury { model outcome splitBps reasoning error } resolvedAt } onChain { vault jobPdaCommit fundingTxSig } proposalDeadline createdAt updatedAt clientProfile { handle cryptoId displayName avatarUrl verified } }
-  }
-}
-"#;
-
-const JOB_QUERY: &str = r#"
-query Job($id: ID!) {
-  job(id: $id) { jobId client title description category skills budget { amount asset chain } status proposalCount groupId contractEscrowId selectedCandidate dispute { reason openedBy openedAt status outcome splitBps judgeModel presided reasoning jury { model outcome splitBps reasoning error } resolvedAt } onChain { vault jobPdaCommit fundingTxSig } proposalDeadline createdAt updatedAt clientProfile { handle cryptoId displayName avatarUrl verified } }
 }
 "#;
