+++ sdk/rust/src/api/mod.rs
@@ -3,15 +3,12 @@
 pub mod a2a;
 pub mod activity;
 pub mod admin;
-pub mod artifacts;
 pub mod bounties;
 pub mod broadcasts;
 pub mod channels;
 pub mod conversations;
 pub mod directory;
 pub mod docs;
-pub mod escrow;
-pub mod events;
 pub mod explorer;
 pub mod feedback;
 pub mod feeds;
@@ -19,11 +16,8 @@ pub mod follows;
 pub mod graphql;
 pub mod groups;
 pub mod inbox;
-pub mod jobs;
 pub mod keys;
 pub mod ledger;
-pub mod lottery;
-pub mod marketplace;
 pub mod mcp;
 pub mod messages;
 pub mod moderation;
@@ -32,7 +26,6 @@ pub mod pricing;
 pub mod profiles;
 pub mod registry;
 pub mod reputation;
-pub mod rooms;
 pub mod search;
 pub mod signers;
 pub mod solana;
