+++ sdk/rust/src/types/graphql.rs
@@ -4,9 +4,7 @@ use std::collections::HashMap;
 
 use serde::{Deserialize, Serialize};
 
-use super::{
-    AgentCard, Identity, JobBudget, JobDispute, JobOnChain, LedgerReference, MarketplacePrice,
-};
+use super::{AgentCard, Identity, LedgerReference};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -139,196 +137,6 @@ pub struct GqlIdentity {
     pub identity: Identity,
     #[serde(skip_serializing_if = "Option::is_none", default)]
     pub owner: Option<GqlProfile>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlProduct {
-    pub product_id: String,
-    pub name: String,
-    pub description: String,
-    pub category: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub tags: Option<Vec<String>>,
-    pub price: MarketplacePrice,
-    pub delivery_method: String,
-    pub status: String,
-    pub sales_count: i64,
-    pub rating: f64,
-    pub created_at: String,
-    pub updated_at: String,
-    pub seller: FeedAuthor,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlProductListResult {
-    pub products: Vec<GqlProduct>,
-    pub count: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentityListing {
-    pub listing_id: String,
-    #[serde(rename = "type")]
-    pub type_: String,
-    pub name: String,
-    pub seller: FeedAuthor,
-    pub seller_crypto_id: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub description: Option<String>,
-    pub category: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub tags: Option<Vec<String>>,
-    pub price: MarketplacePrice,
-    pub listing_type: String,
-    pub status: String,
-    pub created_at: String,
-    pub updated_at: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub expires_at: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub reserve_price: Option<MarketplacePrice>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub highest_bid: Option<GqlIdentityBid>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub winning_bid_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub payment_due_at: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub settlement_status: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentityListingDetail {
-    #[serde(flatten)]
-    pub listing: GqlIdentityListing,
-    pub bids: Vec<GqlIdentityBid>,
-    pub history: Vec<GqlIdentitySale>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentityBid {
-    pub bid_id: String,
-    pub listing_id: String,
-    pub bidder: FeedAuthor,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub bidder_crypto_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub bidder_public_key: Option<String>,
-    pub price: MarketplacePrice,
-    pub status: String,
-    pub created_at: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentityOffer {
-    pub offer_id: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub listing_id: Option<String>,
-    pub name: String,
-    pub buyer: FeedAuthor,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub buyer_crypto_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub buyer_public_key: Option<String>,
-    pub price: MarketplacePrice,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub expires_at: Option<String>,
-    pub status: String,
-    pub created_at: String,
-    pub updated_at: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentitySale {
-    pub sale_id: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub listing_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub offer_id: Option<String>,
-    pub name: String,
-    pub seller: FeedAuthor,
-    pub buyer: FeedAuthor,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub buyer_crypto_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub buyer_public_key: Option<String>,
-    pub price: MarketplacePrice,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub ledger_tx_id: Option<String>,
-    pub created_at: String,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentityListingListResult {
-    pub listings: Vec<GqlIdentityListing>,
-    pub count: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentityBidListResult {
-    pub bids: Vec<GqlIdentityBid>,
-    pub count: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentityOfferListResult {
-    pub offers: Vec<GqlIdentityOffer>,
-    pub count: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlIdentitySaleListResult {
-    pub sales: Vec<GqlIdentitySale>,
-    pub count: i64,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlJobPosting {
-    pub job_id: String,
-    pub client: String,
-    pub title: String,
-    pub description: String,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub category: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub skills: Option<Vec<String>>,
-    pub budget: JobBudget,
-    pub status: String,
-    pub proposal_count: i64,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub group_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub contract_escrow_id: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub selected_candidate: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub dispute: Option<JobDispute>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub on_chain: Option<JobOnChain>,
-    #[serde(skip_serializing_if = "Option::is_none", default)]
-    pub proposal_deadline: Option<String>,
-    pub created_at: String,
-    pub updated_at: String,
-    pub client_profile: FeedAuthor,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct GqlJobListResult {
-    pub jobs: Vec<GqlJobPosting>,
-    pub count: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
