+++ sdk/rust/src/types/mod.rs
@@ -6,26 +6,19 @@
 //! references resolve as `crate::types::SomeType`, like the TS barrel.
 
 mod activity;
-mod artifacts;
 mod bounties;
 mod broadcasts;
 mod commerce;
 mod conversations;
 mod directory;
 mod docs;
-mod escrow;
-mod events;
 mod explorer;
 mod feedback;
 mod follows;
-mod games;
 mod graphql;
 mod groups;
 mod identity;
-mod jobs;
 mod ledger;
-mod lottery;
-mod marketplace;
 mod mcp;
 mod messaging;
 mod payments;
@@ -37,26 +30,19 @@ mod solana;
 mod user;
 
 pub use activity::*;
-pub use artifacts::*;
 pub use bounties::*;
 pub use broadcasts::*;
 pub use commerce::*;
 pub use conversations::*;
 pub use directory::*;
 pub use docs::*;
-pub use escrow::*;
-pub use events::*;
 pub use explorer::*;
 pub use feedback::*;
 pub use follows::*;
-pub use games::*;
 pub use graphql::*;
 pub use groups::*;
 pub use identity::*;
-pub use jobs::*;
 pub use ledger::*;
-pub use lottery::*;
-pub use marketplace::*;
 pub use mcp::*;
 pub use messaging::*;
 pub use payments::*;
