+++ sdk/rust/tests/api_misc.rs
@@ -1,8 +1,8 @@
-//! Endpoint tests for the remaining API namespaces: rooms, lottery, reputation,
-//! explorer, admin, moderation, stats, and artifacts. Each test points the
-//! client at a catch-all mock, invokes one public method, and asserts the
-//! request method + path. Response bodies are permissive — the goal is to
-//! exercise request construction, auth signing, and the response pipeline.
+//! Endpoint tests for the remaining API namespaces: reputation, explorer,
+//! admin, moderation, and stats. Each test points the client at a catch-all
+//! mock, invokes one public method, and asserts the request method + path.
+//! Response bodies are permissive — the goal is to exercise request
+//! construction, auth signing, and the response pipeline.
 
 mod common;
 
@@ -15,290 +15,10 @@ use tinyplace::api::moderation::{
     ModerationActionsQueryParams, ModerationAppealCreate, ModerationStatusUpdate,
 };
 use tinyplace::types::{
-    ArtifactCreateRequest, ArtifactQueryParams, ArtifactRecipientUpdate, AttestationCreate,
-    FeeConfig, FeeResolveParams, FeedbackCreate, FeedbackListParams, FeedbackStatusUpdate,
-    FeedbackVoteRequest, GameActionRequest, GameEmergencyWithdrawalRequest, GameJoinRequest,
-    GameLeaveRequest, GameOperatorRequest, GameRoom, GameRoomQueryParams, GameSettleRequest,
-    LeaderboardQueryParams, LotteryBuyRequest, LotteryDrawRequest, LotteryRoundQueryParams,
-    ModerationAction, ModerationReportCreate, ReputationReviewCreate, ReputationVouchCreate,
-    TrustGraphQueryParams,
+    AttestationCreate, FeeConfig, FeeResolveParams, FeedbackCreate, FeedbackListParams,
+    FeedbackStatusUpdate, FeedbackVoteRequest, LeaderboardQueryParams, ModerationAction,
+    ModerationReportCreate, ReputationReviewCreate, ReputationVouchCreate, TrustGraphQueryParams,
 };
-
-// --- RoomsApi ---
-
-#[tokio::test]
-async fn rooms_list() {
-    let server = any_ok(json!({"rooms": []})).await;
-    let client = client_for(&server);
-    let _ = client
-        .rooms
-        .list(Some(&GameRoomQueryParams::default()))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/rooms"));
-}
-
-#[tokio::test]
-async fn rooms_create() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let room: GameRoom = serde_json::from_value(json!({
-        "roomId": "r1",
-        "game": "poker",
-        "variant": "nlhe",
-        "name": "Table 1",
-        "creator": "@alice",
-        "stakes": {"smallBlind": "1", "bigBlind": "2", "asset": "USDC", "network": "solana"},
-        "buyIn": {"min": "100", "max": "200"},
-        "escrow": {"contract": "c", "network": "solana"},
-        "seats": 6,
-        "players": [],
-        "observerCount": 0,
-        "speed": "normal",
-        "timeouts": {"decision": 30, "disconnectGrace": 60},
-        "rake": {"rate": "0.05", "cap": "10"},
-        "handNumber": 0,
-        "status": "open",
-        "createdAt": "2026-01-01T00:00:00Z",
-        "updatedAt": "2026-01-01T00:00:00Z"
-    }))
-    .unwrap();
-    let _ = client.rooms.create(&room, Some("@alice")).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms"));
-}
-
-#[tokio::test]
-async fn rooms_get() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.rooms.get("r1", None).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/rooms/r1"));
-}
-
-#[tokio::test]
-async fn rooms_join() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .rooms
-        .join("r1", Some(&GameJoinRequest::default()), Some("@alice"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/join"));
-}
-
-#[tokio::test]
-async fn rooms_leave() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .rooms
-        .leave("r1", Some(&GameLeaveRequest::default()), Some("@alice"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/leave"));
-}
-
-#[tokio::test]
-async fn rooms_close() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .rooms
-        .close("r1", Some(&GameOperatorRequest::default()), Some("@op"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/close"));
-}
-
-#[tokio::test]
-async fn rooms_emergency_withdrawal() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let body: GameEmergencyWithdrawalRequest = serde_json::from_value(json!({
-        "agentId": "@alice",
-        "requestTxHash": "0xabc"
-    }))
-    .unwrap();
-    let _ = client
-        .rooms
-        .emergency_withdrawal("r1", &body, Some("@op"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/emergency-withdrawals"));
-}
-
-#[tokio::test]
-async fn rooms_action() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let body: GameActionRequest = serde_json::from_value(json!({
-        "agentId": "@alice",
-        "action": "call"
-    }))
-    .unwrap();
-    let _ = client.rooms.action("r1", &body, Some("@alice")).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/action"));
-}
-
-#[tokio::test]
-async fn rooms_timeout() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .rooms
-        .timeout("r1", Some(&GameOperatorRequest::default()), Some("@op"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/timeout"));
-}
-
-#[tokio::test]
-async fn rooms_list_hands() {
-    let server = any_ok(json!({"hands": []})).await;
-    let client = client_for(&server);
-    let _ = client.rooms.list_hands("r1", None).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/rooms/r1/hands"));
-}
-
-#[tokio::test]
-async fn rooms_start_hand() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .rooms
-        .start_hand("r1", Some(&GameOperatorRequest::default()), Some("@op"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/hands"));
-}
-
-#[tokio::test]
-async fn rooms_get_hand() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.rooms.get_hand("r1", "h1", None).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/rooms/r1/hands/h1"));
-}
-
-#[tokio::test]
-async fn rooms_collusion_report() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.rooms.collusion_report("r1").await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/rooms/r1/collusion"));
-}
-
-#[tokio::test]
-async fn rooms_settle_hand() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let body: GameSettleRequest = serde_json::from_value(json!({
-        "operator": "@op",
-        "winners": [],
-        "txHash": "0xabc"
-    }))
-    .unwrap();
-    let _ = client
-        .rooms
-        .settle_hand("r1", "h1", &body, Some("@op"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/rooms/r1/hands/h1/settle"));
-}
-
-// --- LotteryApi ---
-
-#[tokio::test]
-async fn lottery_current() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.lottery.current(None).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/lottery"));
-}
-
-#[tokio::test]
-async fn lottery_list_rounds() {
-    let server = any_ok(json!({"rounds": []})).await;
-    let client = client_for(&server);
-    let _ = client
-        .lottery
-        .list_rounds(Some(&LotteryRoundQueryParams::default()))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/lottery/rounds"));
-}
-
-#[tokio::test]
-async fn lottery_get_round() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.lottery.get_round("rd1").await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/lottery/rounds/rd1"));
-}
-
-#[tokio::test]
-async fn lottery_holdings() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.lottery.holdings(Some("@alice")).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/lottery/holdings"));
-}
-
-#[tokio::test]
-async fn lottery_buy() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let request: LotteryBuyRequest = serde_json::from_value(json!({
-        "agentId": "@alice",
-        "amountMicros": "1000000"
-    }))
-    .unwrap();
-    let _ = client.lottery.buy(&request, Some("@alice")).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/lottery/buy"));
-}
-
-#[tokio::test]
-async fn lottery_draw() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .lottery
-        .draw("rd1", Some(&LotteryDrawRequest::default()), Some("@op"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/lottery/rounds/rd1/draw"));
-}
 
 // --- ReputationApi ---
 
@@ -952,78 +672,6 @@ async fn stats_fees() {
     assert_eq!(req.method.as_str(), "GET");
     assert!(req.url.path().contains("/stats/fees"));
 }
-
-// --- ArtifactsApi ---
-
-#[tokio::test]
-async fn artifacts_list() {
-    let server = any_ok(json!({"artifacts": []})).await;
-    let client = client_for(&server);
-    let _ = client
-        .artifacts
-        .list(Some(&ArtifactQueryParams::default()), Some("@alice"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/artifacts"));
-}
-
-#[tokio::test]
-async fn artifacts_create() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .artifacts
-        .create(&ArtifactCreateRequest::default(), Some("@alice"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "POST");
-    assert!(req.url.path().contains("/artifacts"));
-}
-
-#[tokio::test]
-async fn artifacts_get() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.artifacts.get("art1", Some("@alice")).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/artifacts/art1"));
-}
-
-#[tokio::test]
-async fn artifacts_remove() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.artifacts.remove("art1", Some("@alice")).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "DELETE");
-    assert!(req.url.path().contains("/artifacts/art1"));
-}
-
-#[tokio::test]
-async fn artifacts_download() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client.artifacts.download("art1", Some("@alice")).await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "GET");
-    assert!(req.url.path().contains("/artifacts/art1/download"));
-}
-
-#[tokio::test]
-async fn artifacts_update_recipients() {
-    let server = any_empty_ok().await;
-    let client = client_for(&server);
-    let _ = client
-        .artifacts
-        .update_recipients("art1", &ArtifactRecipientUpdate::default(), Some("@alice"))
-        .await;
-    let req = only_request(&server).await;
-    assert_eq!(req.method.as_str(), "PUT");
-    assert!(req.url.path().contains("/artifacts/art1/recipients"));
-}
-
 // --- FeedbackApi ---
 
 #[tokio::test]
